Move check_for_loop_arg back to mod; split into 4 lint files
